refactoring sim
