+++ descwl_shear_sims/sim/constants.py
@@ -1,0 +1,10 @@
+ZERO_POINT = 30.0
+SCALE = 0.2  # arcseconds/pixel
+RANDOM_DENSITY = 80  # per square arcmin
+GRID_N_ON_SIDE = 6  # grids are 6x6 on a side
+
+# fwhm if the special fixed psf
+FIXED_PSF_FWHM = 0.8
+
+# for fixed moffat psfs
+FIXED_MOFFAT_BETA = 2.5
