+++ descwl_shear_sims/sim/shifts.py
@@ -1,0 +1,131 @@
+import numpy as np
+
+from .constants import (
+    RANDOM_DENSITY,
+    GRID_N_ON_SIDE,
+    SCALE,
+)
+
+
+def get_shifts(
+    *,
+    rng,
+    coadd_dim,
+    buff,
+    layout,
+    nobj=None,
+):
+    """
+    make position shifts for objects
+
+    rng: numpy.random.RandomState
+        Numpy random state
+    coadd_dim: int
+        Dimensions of final coadd
+    buff: int
+        Buffer region where no objects will be drawn
+    layout: string
+        'grid' or 'random'
+    nobj: int, optional
+        Optional number of objects to draw, defaults to None
+        in which case a poission deviate is draw according
+        to RANDOM_DENSITY
+    """
+
+    if layout == 'grid':
+        shifts = get_grid_shifts(
+            rng=rng,
+            dim=coadd_dim,
+            n_on_side=GRID_N_ON_SIDE,
+        )
+    elif layout == 'random':
+        # area covered by objects
+        if nobj is None:
+            area = ((coadd_dim - 2*buff)*SCALE/60)**2
+            nobj_mean = area * RANDOM_DENSITY
+            nobj = rng.poisson(nobj_mean)
+
+        shifts = get_random_shifts(
+            rng=rng,
+            dim=coadd_dim,
+            buff=buff,
+            size=nobj,
+        )
+    else:
+        raise ValueError("bad layout: '%s'" % layout)
+
+    return shifts
+
+
+def get_grid_shifts(*, rng, dim, n_on_side):
+    """
+    get a set of gridded shifts, with random shifts at the pixel scale
+
+    Parameters
+    ----------
+    rng: numpy.random.RandomState
+        The random number generator
+    dim: int
+        Dimensions of the final image
+    n_on_side: int
+        Number of objects on each side
+
+    Returns
+    -------
+    shifts: array
+        Array with dx, dy offset fields for each point, in
+        arcsec
+    """
+    spacing = dim/(n_on_side+1)*SCALE
+
+    ntot = n_on_side**2
+
+    # ix/iy are really on the sky
+    grid = spacing*(np.arange(n_on_side) - (n_on_side-1)/2)
+
+    shifts = np.zeros(ntot, dtype=[('dx', 'f8'), ('dy', 'f8')])
+
+    i = 0
+    for ix in range(n_on_side):
+        for iy in range(n_on_side):
+            dx = grid[ix] + SCALE*rng.uniform(low=-0.5, high=0.5)
+            dy = grid[iy] + SCALE*rng.uniform(low=-0.5, high=0.5)
+
+            shifts['dx'][i] = dx
+            shifts['dy'][i] = dy
+            i += 1
+
+    return shifts
+
+
+def get_random_shifts(*, rng, dim, buff, size):
+    """
+    get a set of gridded shifts, with random shifts at the pixel scale
+
+    Parameters
+    ----------
+    rng: numpy.random.RandomState
+        The random number generator
+    dim: int
+        Dimensions of the final image
+    n_on_side: int
+        Number of objects on each side
+
+    Returns
+    -------
+    shifts: array
+        Array with dx, dy offset fields for each point, in
+        arcsec
+    """
+
+    halfwidth = (dim - 2*buff)/2.0
+
+    low = -halfwidth*SCALE
+    high = halfwidth*SCALE
+
+    shifts = np.zeros(size, dtype=[('dx', 'f8'), ('dy', 'f8')])
+
+    shifts['dx'] = rng.uniform(low=low, high=high, size=size)
+    shifts['dy'] = rng.uniform(low=low, high=high, size=size)
+
+    return shifts
