+++ descwl_shear_sims/sim/sim.py
@@ -1,13 +1,9 @@
-import os
 from numba import njit
 import copy
 import galsim
 import numpy as np
-import descwl
 from ..se_obs import SEObs
-from ..cache_tools import cached_catalog_read
 from ..ps_psf import PowerSpectrumPSF
-from ..stars import load_sample_stars, sample_star_density, get_star_mag
 from ..saturation import saturate_image_and_mask, BAND_SAT_VALS
 from ..lsst_bits import BAD_COLUMN, COSMIC_RAY, SAT, BRIGHT
 from ..gen_star_masks import add_bright_star_mask
@@ -15,18 +11,19 @@ from ..gen_masks import (
     generate_basic_mask, generate_cosmic_rays, generate_bad_columns,
 )
 from ..star_bleeds import add_bleed, get_max_mag_with_bleed
+from .surveys import get_survey, rescale_wldeblend_images
 
-PSF_FWHM = 0.8
-MOFFAT_BETA = 2.5
-SCALE = 0.2
+from .constants import (
+    SCALE,
+    FIXED_PSF_FWHM,
+    FIXED_MOFFAT_BETA,
+)
+
 WORLD_ORIGIN = galsim.CelestialCoord(
     ra=200 * galsim.degrees,
     dec=0 * galsim.degrees,
 )
 
-GRID_N_ON_SIDE = 6
-RANDOM_DENSITY = 80  # per square arcmin
-ZERO_POINT = 30.0
 
 DEFAULT_SIM_CONFIG = {
     "gal_type": "exp",
@@ -44,11 +41,6 @@ DEFAULT_SIM_CONFIG = {
     "star_bleeds": False,
     "cosmic_rays": False,
     "bad_columns": False,
-}
-
-DEFAULT_FIXED_GAL_CONFIG = {
-    "mag": 17.0,
-    "hlr": 0.5,
 }
 
 
@@ -185,32 +177,6 @@ def make_sim(
     }
 
 
-def get_survey(*, gal_type, band):
-    """
-    Get a survey object
-
-    Parameters
-    ----------
-    gal_type: string
-        'wldeblend' or 'exp'
-    band: string
-        e.g. 'r'
-
-    Returns
-    -------
-    For gal_type 'wldeblend' returns a WLDeblendSurvey object,
-    for gal_type 'exp' returns BasicSurvey.
-    """
-    if gal_type == 'wldeblend':
-        survey = WLDeblendSurvey(band=band)
-    elif gal_type in ['exp']:
-        survey = BasicSurvey(band=band)
-    else:
-        raise ValueError("bad gal_type: '%s'" % gal_type)
-
-    return survey
-
-
 def get_objlist(*, galaxy_catalog, survey, g1, g2, star_catalog=None, noise=None):
     """
     get the objlist and shifts, possibly combining the galaxy catalog
@@ -259,55 +225,6 @@ def get_objlist(*, galaxy_catalog, survey, g1, g2, star_catalog=None, noise=None
         bright_mags = None
 
     return objlist, shifts, bright_objlist, bright_shifts, bright_mags
-
-
-def make_galaxy_catalog(
-    *,
-    rng,
-    gal_type,
-    coadd_dim,
-    buff,
-    layout=None,
-    gal_config=None,
-):
-    """
-    rng: numpy.random.RandomState
-        Numpy random state
-    gal_type: string
-        'exp' or 'wldeblend'
-    coadd_dim: int
-        Dimensions of coadd
-    buff: int
-        Buffer around the edge where no objects are drawn
-    layout: string, optional
-        'grid' or 'random'.  Ignored for gal_type "wldeblend", otherwise
-        required.
-    gal_config: dict or None
-        Can be sent for fixed galaxy catalog.  See DEFAULT_FIXED_GAL_CONFIG
-        for defaults
-    """
-    if gal_type == 'wldeblend':
-        galaxy_catalog = WLDeblendGalaxyCatalog(
-            rng=rng,
-            coadd_dim=coadd_dim,
-            buff=buff,
-        )
-    else:
-
-        if layout is None:
-            raise ValueError("send layout= for gal_type '%s'" % gal_type)
-
-        gal_config = get_fixed_gal_config(config=gal_config)
-        galaxy_catalog = FixedGalaxyCatalog(
-            rng=rng,
-            coadd_dim=coadd_dim,
-            buff=buff,
-            layout=layout,
-            mag=gal_config['mag'],
-            hlr=gal_config['hlr'],
-        )
-
-    return galaxy_catalog
 
 
 def make_wcs(*, scale, image_origin, world_origin, theta=None):
@@ -374,130 +291,6 @@ def make_coadd_wcs(coadd_dim):
         image_origin=coadd_origin,
         world_origin=WORLD_ORIGIN,
     )
-
-
-def get_shifts(
-    *,
-    rng,
-    coadd_dim,
-    buff,
-    layout,
-    nobj=None,
-):
-    """
-    make position shifts for objects
-
-    rng: numpy.random.RandomState
-        Numpy random state
-    coadd_dim: int
-        Dimensions of final coadd
-    buff: int
-        Buffer region where no objects will be drawn
-    layout: string
-        'grid' or 'random'
-    nobj: int, optional
-        Optional number of objects to draw, defaults to None
-        in which case a poission deviate is draw according
-        to RANDOM_DENSITY
-    """
-
-    if layout == 'grid':
-        shifts = get_grid_shifts(
-            rng=rng,
-            dim=coadd_dim,
-            n_on_side=GRID_N_ON_SIDE,
-        )
-    elif layout == 'random':
-        # area covered by objects
-        if nobj is None:
-            area = ((coadd_dim - 2*buff)*SCALE/60)**2
-            nobj_mean = area * RANDOM_DENSITY
-            nobj = rng.poisson(nobj_mean)
-
-        shifts = get_random_shifts(
-            rng=rng,
-            dim=coadd_dim,
-            buff=buff,
-            size=nobj,
-        )
-    else:
-        raise ValueError("bad layout: '%s'" % layout)
-
-    return shifts
-
-
-def get_grid_shifts(*, rng, dim, n_on_side):
-    """
-    get a set of gridded shifts, with random shifts at the pixel scale
-
-    Parameters
-    ----------
-    rng: numpy.random.RandomState
-        The random number generator
-    dim: int
-        Dimensions of the final image
-    n_on_side: int
-        Number of objects on each side
-
-    Returns
-    -------
-    shifts: array
-        Array with dx, dy offset fields for each point, in
-        arcsec
-    """
-    spacing = dim/(n_on_side+1)*SCALE
-
-    ntot = n_on_side**2
-
-    # ix/iy are really on the sky
-    grid = spacing*(np.arange(n_on_side) - (n_on_side-1)/2)
-
-    shifts = np.zeros(ntot, dtype=[('dx', 'f8'), ('dy', 'f8')])
-
-    i = 0
-    for ix in range(n_on_side):
-        for iy in range(n_on_side):
-            dx = grid[ix] + SCALE*rng.uniform(low=-0.5, high=0.5)
-            dy = grid[iy] + SCALE*rng.uniform(low=-0.5, high=0.5)
-
-            shifts['dx'][i] = dx
-            shifts['dy'][i] = dy
-            i += 1
-
-    return shifts
-
-
-def get_random_shifts(*, rng, dim, buff, size):
-    """
-    get a set of gridded shifts, with random shifts at the pixel scale
-
-    Parameters
-    ----------
-    rng: numpy.random.RandomState
-        The random number generator
-    dim: int
-        Dimensions of the final image
-    n_on_side: int
-        Number of objects on each side
-
-    Returns
-    -------
-    shifts: array
-        Array with dx, dy offset fields for each point, in
-        arcsec
-    """
-
-    halfwidth = (dim - 2*buff)/2.0
-
-    low = -halfwidth*SCALE
-    high = halfwidth*SCALE
-
-    shifts = np.zeros(size, dtype=[('dx', 'f8'), ('dy', 'f8')])
-
-    shifts['dx'] = rng.uniform(low=low, high=high, size=size)
-    shifts['dy'] = rng.uniform(low=low, high=high, size=size)
-
-    return shifts
 
 
 class FixedPSF(object):
@@ -869,51 +662,6 @@ def get_convolved_objects(*, objlist, psf, shifts, offset, se_wcs, se_origin):
     return convolved_objects, psf_gsobj
 
 
-def rescale_wldeblend_images(*, survey, image, noise, weight):
-    """
-    Rescale wldeblend images noise and weight to our zero point
-
-    Parameters
-    ----------
-    survey: WLDeblendSurvey
-        The survey object
-    image: galsim.Image or ndarray
-        The image to be scaled
-    noise: galsim.Image or ndarray
-        The noise image to be scaled
-    weight: galsim.Image or ndarray
-        The weightimage to be scaled
-
-    Returns
-    -------
-    None
-    """
-    fac = get_wldeblend_rescale_fac(survey)
-    wfac = 1.0/fac**2
-
-    image *= fac
-    noise *= fac
-    weight *= wfac
-
-
-def get_wldeblend_rescale_fac(survey):
-    """
-    Get the factor to rescale wldeblend images to our zero point
-
-    Parameters
-    ----------
-    survey: WLDeblendSurvey
-        The survey object
-
-    Returns
-    -------
-    number by which to rescale images
-    """
-    s_zp = survey.zero_point
-    s_et = survey.exposure_time
-    return 10.0**(0.4*(ZERO_POINT - 24.0))/s_zp/s_et
-
-
 def get_convolved_objlist_variable_psf(
     *,
     objlist,
@@ -1081,518 +829,6 @@ def get_sim_config(config=None):
     return out_config
 
 
-def get_fixed_gal_config(config=None):
-    """
-    get the configuration for fixed galaxies, with defaults in place
-
-    Parameters
-    ----------
-    config: dict, optional
-        The input config. Over-rides defaults
-
-    Returns
-    -------
-    the config dict
-    """
-    out_config = copy.deepcopy(DEFAULT_FIXED_GAL_CONFIG)
-
-    if config is not None:
-        for key in config:
-            if key not in out_config:
-                raise ValueError("bad key for fixed gals: '%s'" % key)
-        out_config.update(config)
-    return out_config
-
-
-class WLDeblendSurvey(object):
-    """
-    wrapper for wldeblend surveys
-
-    Parameters
-    ----------
-    band: string
-        The band, e.g. 'r'
-    """
-    def __init__(self, *, band):
-
-        pars = descwl.survey.Survey.get_defaults(
-            survey_name='LSST',
-            filter_band=band,
-        )
-
-        pars['survey_name'] = 'LSST'
-        pars['filter_band'] = band
-        pars['pixel_scale'] = SCALE
-
-        # note in the way we call the descwl package, the image width
-        # and height is not actually used
-        pars['image_width'] = 10
-        pars['image_height'] = 10
-
-        # some versions take in the PSF and will complain if it is not
-        # given
-        try:
-            svy = descwl.survey.Survey(**pars)
-        except Exception:
-            pars['psf_model'] = None
-            svy = descwl.survey.Survey(**pars)
-
-        self.noise = np.sqrt(svy.mean_sky_level)
-
-        self.descwl_survey = svy
-
-    @property
-    def filter_band(self):
-        """
-        get the filter band for this survey
-
-        Returns
-        -------
-        string filter band, e.g. 'r'
-        """
-        return self.descwl_survey.filter_band
-
-    def get_flux(self, mag):
-        """
-        convert mag to flux
-        """
-        return self.descwl_survey.get_flux(mag)
-
-
-class BasicSurvey(object):
-    """
-    represent a simple survey with common interface
-
-    Parameters
-    ----------
-    band: string
-        e.g. 'r'
-    """
-    def __init__(self, *, band):
-        self.band = band
-        self.noise = 1.0
-        self.filter_band = band
-
-    def get_flux(self, mag):
-        """
-        get the flux for the input mag using the standard zero point
-        """
-        return 10**(0.4 * (ZERO_POINT - mag))
-
-
-class FixedGalaxyCatalog(object):
-    """
-    Galaxies of fixed galsim type, flux, and size
-
-    Same for all bands
-
-    Parameters
-    ----------
-    rng: np.random.RandomState
-        The random number generator
-    coadd_dim: int
-        dimensions of the coadd
-    buff: int
-        Buffer region with no objects, on all sides of image
-    layout: string
-        The layout of objects, either 'grid' or 'random'
-    mag: float
-        Magnitude of all objects
-    hlr: float
-        Half light radius of all objects
-    """
-    def __init__(self, *, rng, coadd_dim, buff, layout, mag, hlr):
-        self.gal_type = 'exp'
-        self.mag = mag
-        self.hlr = hlr
-        self.rng = rng
-
-        self.shifts = get_shifts(
-            rng=rng,
-            coadd_dim=coadd_dim,
-            buff=buff,
-            layout=layout,
-        )
-
-    def get_objlist(self, *, survey, g1, g2):
-        """
-        get a list of galsim objects
-
-        Parameters
-        ----------
-        band: string
-            Get objects for this band.  For the fixed
-            catalog, the objects are the same for every band
-
-        Returns
-        -------
-        [galsim objects]
-        """
-
-        flux = survey.get_flux(self.mag)
-
-        num = self.shifts.size
-        objlist = [
-            self._get_galaxy(i, flux).shear(g1=g1, g2=g2)
-            for i in range(num)
-        ]
-
-        shifts = self.shifts.copy()
-        return objlist, shifts
-
-    def _get_galaxy(self, i, flux):
-        """
-        get a galaxy object
-
-        Parameters
-        ----------
-        i: int
-            Index of object
-        flux: float
-            Flux of object
-
-        Returns
-        --------
-        galsim.GSObject
-        """
-        return galsim.Exponential(
-            half_light_radius=self.hlr,
-            flux=flux,
-        ).shift(
-            dx=self.shifts['dx'][i],
-            dy=self.shifts['dy'][i]
-        )
-
-
-class WLDeblendGalaxyCatalog(object):
-    """
-    Catalog of galaxies from wldeblend
-
-    Parameters
-    ----------
-    rng: np.random.RandomState
-        The random number generator
-    coadd_dim: int
-        Dimensions of the coadd
-    buff: int
-        Buffer region with no objects, on all sides of image
-    """
-    def __init__(self, *, rng, coadd_dim, buff):
-        self.gal_type = 'wldeblend'
-        self.rng = rng
-
-        self._wldeblend_cat = read_wldeblend_cat(rng)
-
-        # one square degree catalog, convert to arcmin
-        gal_dens = self._wldeblend_cat.size / (60 * 60)
-        area = ((coadd_dim - 2*buff)*SCALE/60)**2
-        nobj_mean = area * gal_dens
-        nobj = rng.poisson(nobj_mean)
-
-        self.shifts = get_shifts(
-            rng=rng,
-            coadd_dim=coadd_dim,
-            buff=buff,
-            layout="random",
-            nobj=nobj,
-        )
-
-        num = self.shifts.size
-        self.indices = self.rng.randint(
-            0,
-            self._wldeblend_cat.size,
-            size=num,
-        )
-
-        self.angles = self.rng.uniform(low=0, high=360, size=num)
-
-    def get_objlist(self, *, survey, g1, g2):
-        """
-        get a list of galsim objects
-
-        Parameters
-        ----------
-        survey: WLDeblendSurvey
-            The survey object
-        g1: float
-            The g1 shear to apply to these objects
-        g2: float
-            The g2 shear to apply to these objects
-
-        Returns
-        -------
-        [galsim objects]
-        """
-
-        builder = descwl.model.GalaxyBuilder(
-            survey=survey.descwl_survey,
-            no_disk=False,
-            no_bulge=False,
-            no_agn=False,
-            verbose_model=False,
-        )
-
-        num = self.shifts.size
-
-        band = survey.filter_band
-
-        # object is already shifted, so this results in the scene
-        # being sheared
-        objlist = [
-            self._get_galaxy(builder, band, i).shear(g1=g1, g2=g2)
-            for i in range(num)
-        ]
-
-        shifts = self.shifts.copy()
-        return objlist, shifts
-
-    def _get_galaxy(self, builder, band, i):
-        """
-        Get a galaxy
-
-        Parameters
-        ----------
-        builder: descwl.model.GalaxyBuilder
-            Builder for this object
-        band: string
-            Band string, e.g. 'r'
-        i: int
-            Index of object
-
-        Returns
-        -------
-        galsim.GSObject
-        """
-        index = self.indices[i]
-        dx = self.shifts['dx'][i]
-        dy = self.shifts['dy'][i]
-
-        angle = self.angles[i]
-
-        galaxy = builder.from_catalog(
-            self._wldeblend_cat[index],
-            0,
-            0,
-            band,
-        ).model.rotate(
-            angle * galsim.degrees,
-        ).shift(
-            dx=dx,
-            dy=dy,
-        )
-
-        return galaxy
-
-
-def read_wldeblend_cat(rng):
-    """
-    Read the catalog from the cache, but update the position angles each time
-
-    Parameters
-    ----------
-    rng: np.random.RandomState
-        The random number generator
-
-    Returns
-    -------
-    array with fields
-    """
-    fname = os.path.join(
-        os.environ.get('CATSIM_DIR', '.'),
-        'OneDegSq.fits',
-    )
-
-    # not thread safe
-    cat = cached_catalog_read(fname)
-
-    cat['pa_disk'] = rng.uniform(
-        low=0.0,
-        high=360.0,
-        size=cat.size,
-    )
-    cat['pa_bulge'] = cat['pa_disk']
-    return cat
-
-
-class StarCatalog(object):
-    """
-    Star catalog with variable density
-
-    Parameters
-    ----------
-    rng: np.random.RandomState
-        The random number generator
-    coadd_dim: int
-        Dimensions of the coadd
-    buff: int
-        Buffer region with no objects, on all sides of image
-    density: float, optional
-        Optional density for catalog, if not sent the density is variable and
-        drawn from the expected galactic density
-    """
-    def __init__(self, *, rng, coadd_dim, buff, density=None):
-        self.rng = rng
-
-        self._star_cat = load_sample_stars()
-
-        if density is None:
-            density_mean = sample_star_density(
-                rng=self.rng,
-                min_density=2,
-                max_density=100,
-            )
-        else:
-            density_mean = density
-
-        # one square degree catalog, convert to arcmin
-        area = ((coadd_dim - 2*buff)*SCALE/60)**2
-        nobj_mean = area * density_mean
-        nobj = rng.poisson(nobj_mean)
-
-        self.density = nobj/area
-
-        self.shifts = get_shifts(
-            rng=rng,
-            coadd_dim=coadd_dim,
-            buff=buff,
-            layout="random",
-            nobj=nobj,
-        )
-
-        num = self.shifts.size
-        self.indices = self.rng.randint(
-            0,
-            self._star_cat.size,
-            size=num,
-        )
-
-    def get_objlist(self, *, survey, noise):
-        """
-        get a list of galsim objects
-
-        Parameters
-        ----------
-        survey: WLDeblendSurvey or BasicSurvey
-            The survey object
-        noise: float
-            The noise level, needed for setting gsparams
-
-        Returns
-        -------
-        [galsim objects]
-        """
-
-        num = self.shifts.size
-
-        band = survey.filter_band
-        objlist = []
-        shift_ind = []
-        bright_objlist = []
-        bright_shift_ind = []
-        bright_mags = []
-        for i in range(num):
-            star, mag, isbright = self._get_star(survey, band, i, noise)
-            if isbright:
-                bright_objlist.append(star)
-                bright_shift_ind.append(i)
-                bright_mags.append(mag)
-            else:
-                objlist.append(star)
-                shift_ind.append(i)
-
-        # objlist = [
-        #     self._get_star(survey, band, i, noise)
-        #     for i in range(num)
-        # ]
-
-        shifts = self.shifts[shift_ind].copy()
-        bright_shifts = self.shifts[bright_shift_ind].copy()
-        return objlist, shifts, bright_objlist, bright_shifts, bright_mags
-
-    def _get_star(self, survey, band, i, noise):
-        """
-        Parameters
-        ----------
-        survey: WLDeblendSurvey or BasicSurvey
-            The survey object
-        band: string
-            Band string, e.g. 'r'
-        i: int
-            Index of object
-        noise: float
-            The noise level, needed for setting gsparams
-
-        Returns
-        -------
-        galsim.GSObject
-        """
-
-        index = self.indices[i]
-        dx = self.shifts['dx'][i]
-        dy = self.shifts['dy'][i]
-
-        mag = get_star_mag(stars=self._star_cat, index=index, band=band)
-        flux = survey.get_flux(mag)
-
-        gsparams, isbright = get_star_gsparams(mag, flux, noise)
-        star = galsim.Gaussian(
-            fwhm=1.0e-4,
-            flux=flux,
-            gsparams=gsparams,
-        ).shift(
-            dx=dx,
-            dy=dy,
-        )
-
-        return star, mag, isbright
-
-
-def get_star_gsparams(mag, flux, noise):
-    """
-    Get appropriate gsparams given flux and noise
-
-    Parameters
-    ----------
-    mag: float
-        mag of star
-    flux: float
-        flux of star
-    noise: float
-        noise of image
-
-    Returns
-    --------
-    GSParams, isbright where isbright is true for stars with mag less than 18
-    """
-    do_thresh = do_acc = False
-    if mag < 18:
-        do_thresh = True
-    if mag < 15:
-        do_acc = True
-
-    if do_thresh or do_acc:
-        isbright = True
-
-        kw = {}
-        if do_thresh:
-            folding_threshold = noise/flux
-            folding_threshold = np.exp(
-                np.floor(np.log(folding_threshold))
-            )
-            kw['folding_threshold'] = min(folding_threshold, 0.005)
-
-        if do_acc:
-            kw['kvalue_accuracy'] = 1.0e-8
-            kw['maxk_threshold'] = 1.0e-5
-
-        gsparams = galsim.GSParams(**kw)
-    else:
-        gsparams = None
-        isbright = False
-
-    return gsparams, isbright
-
-
 def make_psf(*, psf_type):
     """
     Make a fixed PSF
@@ -1607,9 +843,9 @@ def make_psf(*, psf_type):
     Gaussian or Moffat
     """
     if psf_type == "gauss":
-        psf = galsim.Gaussian(fwhm=PSF_FWHM)
+        psf = galsim.Gaussian(fwhm=FIXED_PSF_FWHM)
     elif psf_type == "moffat":
-        psf = galsim.Moffat(fwhm=PSF_FWHM, beta=MOFFAT_BETA)
+        psf = galsim.Moffat(fwhm=FIXED_PSF_FWHM, beta=FIXED_MOFFAT_BETA)
     else:
         raise ValueError("bad psf_type '%s'" % psf_type)
